Started Hero Sliders
added the code for hero sliders... fully functinal right now but does not look even close to what I want the finished result to be.
first, make images for the sliders themselves.
Second, change the LOFI Image to a procreate drawn one.
Third, Adjust the height and starting position of the sliders.
Fourth, make sure the images are object fit cover, so that they are scalable/responsive.